requirements.txt
torch>=1.1.0
nltk==0.3.4
numpy>=0.13
scipy>=1.2.0
scikit-learn>=0.17
spacy
matplotlib
tqdm
bokeh
numba
gensim
glove_python
https://github.com/facebookresearch/pysparnn/archive/master.zip
